Modify: Ajustes para un optimo funcionamiento tanto local como en docker. (funcionando)

+++ backend/domain/repositories/AttendeeRepository.js
@@ -1,20 +1,36 @@
-const db = require('../../infrastructure/database/database');
+const { createConnection } = require('../../infrastructure/database/database');
 
 class AttendeeRepository {
-
+    constructor() {
+        this.connection = null;
+    }
+    async connectdb() {
+        try {
+            this.connection = await createConnection();
+        } catch (error) {
+            console.error('Error en connectdb:', error);
+            throw { status: 500, body: { message: 'Error al conectar con la base de datos' } };
+        }
+    }
     async getAllAttendees() {
+        if (!this.connection) {
+            throw { status: 500, body: { message: 'Error al obtener asistentes repository' } };
+        }
         try {
             const query = 'SELECT * FROM attendees';
-            const [rows] = await db.execute(query);
+            const [rows] = await this.connection.execute(query);
             return rows.length > 0 ? rows : null;
         } catch (error) {
             throw error;
         }
     }
     async getAttendeeById(id) {
+        if (!this.connection) {
+            throw { status: 500, body: { message: 'Error al obtener asistente repository' } };
+        }
         try {
             const query = 'SELECT * FROM attendees WHERE id = ?';
-            const [rows] = await db.execute(query, [id]);
+            const [rows] = await this.connection.execute(query, [id]);
             return rows[0];
         } catch (error) {
             throw error;
@@ -22,9 +38,12 @@ class AttendeeRepository {
     }
 
     async createAttendee(attendee) {
+        if (!this.connection) {
+            throw { status: 500, body: { message: 'Error al crear asistente repository' } };
+        }
         try {
             const query = 'INSERT INTO attendees (event_id, user_id) VALUES (?, ?)';
-            const [rows] = await db.execute(query, [attendee.eventId, attendee.userId]);
+            const [rows] = await this.connection.execute(query, [attendee.eventId, attendee.userId]);
             return rows.insertId;
         } catch (error) {
             throw error;
@@ -32,9 +51,12 @@ class AttendeeRepository {
     }
 
     async deleteAttendee(id) {
+        if (!this.connection) {
+            throw { status: 500, body: { message: 'Error al eliminar asistente repository' } };
+        }
         try {
             const query = 'DELETE FROM attendees WHERE id = ?';
-            const [rows] = await db.execute(query, [id]);
+            const [rows] = await this.connection.execute(query, [id]);
             return rows.affectedRows > 0;
         } catch (error) {
             throw error;
@@ -42,9 +64,12 @@ class AttendeeRepository {
     }
 
     async getAttendeesByEventId(eventId) {
+        if (!this.connection) {
+            throw { status: 500, body: { message: 'Error al obtener asistentes repository' } };
+        }
         try {
             const query = 'SELECT * FROM attendees WHERE event_id = ?';
-            const [rows] = await db.execute(query, [eventId]);
+            const [rows] = await this.connection.execute(query, [eventId]);
             return rows.length > 0 ? rows : null;
         } catch (error) {
             throw error;
@@ -52,9 +77,12 @@ class AttendeeRepository {
     }
 
     async getAttendeesByUserId(userId) {
+        if (!this.connection) {
+            throw { status: 500, body: { message: 'Error al obtener asistentes repository' } };
+        }
         try {
             const query = 'SELECT * FROM attendees WHERE user_id = ?';
-            const [rows] = await db.execute(query, [userId]);
+            const [rows] = await this.connection.execute(query, [userId]);
             return rows;
         } catch (error) {
             throw error;
@@ -62,9 +90,12 @@ class AttendeeRepository {
     }
 
     async getAttendeeByEmail(email) {
+        if (!this.connection) {
+            throw { status: 500, body: { message: 'Error al obtener asistentes repository' } };
+        }
         try {
             const query = 'SELECT * FROM attendees WHERE email = ?';
-            const [rows] = await db.execute(query, [email]);
+            const [rows] = await this.connection.execute(query, [email]);
             return rows[0];
         } catch (error) {
             throw error;
@@ -72,9 +103,12 @@ class AttendeeRepository {
     }
 
     async getAttendeeByEmailAndEventId(email, eventId) {
+        if (!this.connection) {
+            throw { status: 500, body: { message: 'Error al obtener asistentes repository' } };
+        }
         try {
             const query = 'SELECT * FROM attendees WHERE email = ? AND event_id = ?';
-            const [rows] = await db.execute(query, [email, eventId]);
+            const [rows] = await this.connection.execute(query, [email, eventId]);
             return rows[0];
         } catch (error) {
             throw error;
@@ -82,9 +116,12 @@ class AttendeeRepository {
     }
 
     async registerAttendee(attendee) {
+        if (!this.connection) {
+            throw { status: 500, body: { message: 'Error al registrar asistentes repository' } };
+        }
         try {
             const query = 'INSERT INTO attendees (name, email, event_id, user_id) VALUES (?, ?, ?, ?)';
-            const [rows] = await db.execute(query, [attendee.name, attendee.email, attendee.eventId, attendee.userId]);
+            const [rows] = await this.connection.execute(query, [attendee.name, attendee.email, attendee.eventId, attendee.userId]);
             return rows.insertId;
         } catch (error) {
             throw error;
@@ -92,9 +129,12 @@ class AttendeeRepository {
     }
 
     async getEventById(id) {
+        if (!this.connection) {
+            throw { status: 500, body: { message: 'Error al obtener eventos repository' } };
+        }
         try {
             const query = 'SELECT * FROM events WHERE id = ?';
-            const [rows] = await db.execute(query, [id]);
+            const [rows] = await this.connection.execute(query, [id]);
             return rows.length > 0 ? rows[0] : null;
         } catch (error) {
             throw error;
@@ -102,9 +142,12 @@ class AttendeeRepository {
     }
 
     async getUserById(id) {
+        if (!this.connection) {
+            throw { status: 500, body: { message: 'Error al obtener usuario repository' } };
+        }
         try {
             const query = 'SELECT * FROM users WHERE id = ?';
-            const [rows] = await db.execute(query, [id]);
+            const [rows] = await this.connection.execute(query, [id]);
             return rows.length > 0 ? rows[0] : null;
         } catch (error) {
             throw error;
